+++ tla-import/src/test/scala/at/forsyte/apalache/io/tlc/TestTlcConfigParser.scala
@@ -59,6 +59,21 @@ class TestTlcConfigParser extends FunSuite {
     assert(config.constReplacements.isEmpty)
   }
 
+  test("CONSTANT assignments with numbers") {
+    val text =
+      """
+        |CONSTANT
+        |N = 10
+        |K = -20
+        |INIT Init
+        |NEXT Next
+      """.stripMargin
+
+    val config = TlcConfigParser(text)
+    assert(config.constAssignments == Map("N" -> "10", "K" -> "-20"))
+    assert(config.constReplacements.isEmpty)
+  }
+
   test("CONSTANT assignments and SYMMETRY") {
     val text =
       """
@@ -121,6 +136,23 @@ class TestTlcConfigParser extends FunSuite {
     assert(config.behaviorSpec == InitNextSpec("Init", "Next"))
     assert(config.invariants == List("Inv1", "Inv2"))
   }
+
+// TODO: should this test succeed or fail?
+//  currently it fails, because the order of TLC config sections is strictly prescribed
+//
+//  test("INIT-NEXT and INVARIANTS reordered") {
+//    val text =
+//      """
+//        |INVARIANT Inv1
+//        |INIT Init
+//        |INVARIANTS Inv2
+//        |NEXT Next
+//      """.stripMargin
+//
+//    val config = TlcConfigParser(text)
+//    assert(config.behaviorSpec == InitNextSpec("Init", "Next"))
+//    assert(config.invariants == List("Inv1", "Inv2"))
+//  }
 
   test("INIT-NEXT and PROPERTIES") {
     val text =